refactor: creating menu, but need fix, remove dir logic and change helper, utils

+++ models/order.go
@@ -15,6 +15,6 @@ type Order struct {
 }
 
 type OrderItem struct {
-	ProductID string `json:"product_id"`
-	Quantity  int    `json:"quantity"`
+	ProductID string  `json:"product_id"` // тип поменять на float64
+	Quantity  float64 `json:"quantity"`
 }
